Enables UUID as primary key for users and tsks
NOTE: I have researching for a way to migrate current data and associations from
integer to uuid. Unfortunately any of solutions found works fine, so I need to
drop the database and recreate it with the new structure.

Not a big deal since there is no customers data in production.